requirements.txt
asgiref==3.8.1
Django==5.2
django-ipware==7.0.1
django-recaptcha==4.1.0
pillow==11.2.1
python-ipware==3.0.0
sqlparse==0.5.3
typing_extensions==4.13.2
